Move model instance creation helpers to a separate file
